pyproject.toml
[tool.poetry]
name = "order-otus"
version = "0.1.0"
description = ""
authors = ["Aleksandr Mazurenko <amazurenko@rvision.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.1.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
